+++ src/core/client/systems/animations.ts
@@ -62,6 +62,5 @@ export async function playAnimation(
         return;
     }
 
-    alt.log(`played the stupid fucking animation`);
     native.taskPlayAnim(alt.Player.local.scriptID, dict, name, 8.0, -1, duration, flags, 0, false, false, false);
 }