+++ src/core/server/athena/main.ts
@@ -70,5 +70,6 @@ export const DEFAULT_CONFIG = {
         'CLEARING'
     ],
     // Vehicle Properties
-    MAXIMUM_FUEL: 100
+    TIME_BETWEEN_VEHICLE_UPDATES: 10000, // 10s
+    FUEL_LOSS_PER_PLAYER_TICK: 25 // Happens every 10s
 };