+++ src/core/server/extensions/Vehicle.ts
@@ -1,10 +1,13 @@
 import * as alt from 'alt-server';
-import { Vehicle_Lock_State } from '../../shared/enums/vehicle';
+import { Vehicle_Behavior, Vehicle_Lock_State } from '../../shared/enums/vehicle';
 import getter from './vehicleFunctions/getter';
 import keys from './vehicleFunctions/keys';
+import newRef from './vehicleFunctions/new';
+import save from './vehicleFunctions/save';
 import setter from './vehicleFunctions/setter';
 import toggle from './vehicleFunctions/toggle';
 import utility from './vehicleFunctions/utility';
+import * as IVeh from '../../shared/interfaces/Vehicle';
 
 /**
  * Overwrites the default functionality of vehicles.
@@ -15,9 +18,13 @@ declare module 'alt-server' {
     export interface Vehicle {
         athenaLockState: Vehicle_Lock_State;
         engineStatus: boolean;
-        keys: Array<string>;
-        owner: number;
-        fuel: number;
+        keys?: Array<string>;
+        fuel?: number;
+        player_id?: number;
+        behavior?: Vehicle_Behavior;
+        data?: Partial<IVeh.Vehicle>; // Special Vehicle Information
+        nextSave?: number;
+        nextUpdate?: number;
     }
 }
 
@@ -28,7 +35,9 @@ export default function onLoad() {
 export const vehicleFuncs = {
     getter,
     setter,
+    new: newRef,
     keys,
+    save,
     toggle,
     utility
 };