+++ src/core/server/extensions/vehicleFunctions/getter.ts
@@ -1,5 +1,6 @@
 import * as alt from 'alt-server';
-import { Vehicle_Lock_State } from '../../../shared/enums/vehicle';
+import { Vehicle_Behavior, Vehicle_Lock_State } from '../../../shared/enums/vehicle';
+import { isFlagEnabled } from '../../../shared/utility/flags';
 
 function lockState(v: alt.Vehicle): Vehicle_Lock_State {
     if (v.athenaLockState === null || v.athenaLockState === undefined) {
@@ -9,23 +10,29 @@ function lockState(v: alt.Vehicle): Vehicle_Lock_State {
     return v.athenaLockState;
 }
 
-function isOwner(v: alt.Vehicle, target: alt.Player | number): boolean {
-    // If the vehicle has a null owner. There is no owner.
-    if (!v.owner) {
+function hasFuel(v: alt.Vehicle): boolean {
+    if (isFlagEnabled(v.behavior, Vehicle_Behavior.UNLIMITED_FUEL)) {
         return true;
     }
 
-    if (target instanceof alt.Player) {
-        if (v.owner !== target.id) {
-            return false;
-        }
-    } else {
-        if (v.owner !== target) {
-            return false;
-        }
+    if (v.fuel <= 0) {
+        return false;
     }
 
     return true;
 }
 
-export default { isOwner, lockState };
+function isOwner(v: alt.Vehicle, target: alt.Player): boolean {
+    // If the vehicle has a null owner. There is no owner.
+    if (v.player_id === null || v.player_id === undefined) {
+        return true;
+    }
+
+    if (v.player_id !== target.id) {
+        return false;
+    }
+
+    return true;
+}
+
+export default { isOwner, lockState, hasFuel };