+++ src/core/server/extensions/vehicleFunctions/toggle.ts
@@ -1,5 +1,11 @@
 import * as alt from 'alt-server';
-import { Vehicle_Lock_State, Vehicle_Lock_States, Vehicle_State } from '../../../shared/enums/vehicle';
+import {
+    Vehicle_Behavior,
+    Vehicle_Lock_State,
+    Vehicle_Lock_States,
+    Vehicle_State
+} from '../../../shared/enums/vehicle';
+import { isFlagEnabled } from '../../../shared/utility/flags';
 import { playerFuncs } from '../Player';
 import getter from './getter';
 import keys from './keys';
@@ -41,14 +47,26 @@ function lock(v: alt.Vehicle, player: alt.Player, bypass: boolean = false): Vehi
     return v.athenaLockState;
 }
 
-function engine(v: alt.Vehicle, player: alt.Player): void {
-    if (!getter.isOwner(v, player) && !keys.has(v, player)) {
+function engine(v: alt.Vehicle, player: alt.Player, bypass: boolean = false): void {
+    if (isFlagEnabled(v.behavior, Vehicle_Behavior.NEED_KEY_TO_START) && !bypass) {
+        if (!getter.isOwner(v, player) && !keys.has(v, player)) {
+            return;
+        }
+    }
+
+    if (!getter.hasFuel(v)) {
+        v.engineStatus = false;
+        v.setStreamSyncedMeta(Vehicle_State.ENGINE, v.engineStatus);
+        playerFuncs.emit.notification(player, `~r~No Fuel`);
         return;
     }
 
-    v.engineStatus = !v.engineStatus ? true : false;
+    v.engineStatus = !v.engineStatus;
     v.setStreamSyncedMeta(Vehicle_State.ENGINE, v.engineStatus);
-    playerFuncs.emit.notification(player, `Engine ~y~${v.engineStatus ? 'On' : 'Off'}`);
+
+    if (player) {
+        playerFuncs.emit.notification(player, `Engine ~y~${v.engineStatus ? 'On' : 'Off'}`);
+    }
 }
 
 export default {