+++ src/core/server/systems/tick.ts
@@ -2,11 +2,12 @@ import * as alt from 'alt-server';
 import { SYSTEM_EVENTS } from '../../shared/enums/system';
 import { DEFAULT_CONFIG } from '../athena/main';
 import { playerFuncs } from '../extensions/Player';
+import { vehicleFuncs } from '../extensions/Vehicle';
 import { InventoryController } from '../views/inventory';
 import './blip';
 import './interaction';
-import './vehicle';
 import './toolbar';
+import './vehicle';
 
 const timeBetweenPings = 4950;
 
@@ -48,5 +49,20 @@ function handlePing(player: alt.Player): void {
         player.nextFoodSync = Date.now() + DEFAULT_CONFIG.TIME_BETWEEN_FOOD_UPDATES;
         playerFuncs.sync.food(player);
         playerFuncs.sync.water(player);
+    }
+
+    if (player.vehicle) {
+        if (!player.vehicle.nextUpdate || Date.now() > player.vehicle.nextUpdate) {
+            player.vehicle.nextUpdate = Date.now() + DEFAULT_CONFIG.TIME_BETWEEN_VEHICLE_UPDATES;
+            vehicleFuncs.setter.updateFuel(player.vehicle);
+        }
+
+        if (!player.vehicle.nextSave || Date.now() > player.vehicle.nextSave) {
+            player.vehicle.nextSave = Date.now() + Math.floor(Math.random() * 30000) + 10000;
+            const owner = alt.Player.all.find((p) => p.id === player.vehicle.player_id);
+            if (owner) {
+                vehicleFuncs.save.data(player, player.vehicle);
+            }
+        }
     }
 }